requirements.txt
fastapi
httpx
uvicorn[standard]
gunicorn
azure-mcp
